Refactoring gestione utente v2

+++ app/GestioneUtente/gestore_utente.py
@@ -53,7 +53,7 @@ class GestoreUtente:
         cognome=dati_form.get('cognome')
         numTelefono=dati_form.get('numTelefono')
 
-        if nome=="" or cognome =="" or email =="" or ruolo =="" or numTelefono=="":
+        if not nome or not cognome or not email or not ruolo:
             raise ValueError("Compila tutti i campi")
 
         # Controllo unicità email
@@ -64,7 +64,7 @@ class GestoreUtente:
         if not GestoreUtente.validaPassword(password):
             raise ValueError("La password deve avere almeno 8 caratteri, un numero e un carattere speciale.")
 
-        # Hashing
+        # Faccio l'hashing della password
         hashed_pw = generate_password_hash(password)
 
         if ruolo == 'studente':
@@ -72,29 +72,13 @@ class GestoreUtente:
             facolta = dati_form.get('facolta')
             universita = dati_form.get('universita')
 
-            if corso=="" or facolta =="" or universita =="":
+            if not corso or not facolta  or not universita:
                 raise ValueError("Compila tutti i campi")
 
-            nuovo_utente = Studente(
-                nome=nome,
-                cognome=cognome,
-                email=email,
-                numTelefono=numTelefono,
-                password=hashed_pw,
-                verificato=False, 
-                corso=corso,
-                facolta=facolta,
-                universita=universita
-            )
+            nuovo_utente = Studente( nome=nome, cognome=cognome,email=email,numTelefono=numTelefono,password=hashed_pw,verificato=False, 
+                                    corso=corso,facolta=facolta,universita=universita)
         elif ruolo == 'locatore':
-            nuovo_utente = Locatore(
-                nome=nome,
-                cognome=cognome,
-                email=email,
-                numTelefono=numTelefono,
-                password=hashed_pw,
-                verificato=False
-            )
+            nuovo_utente = Locatore( nome=nome, cognome=cognome, email=email, numTelefono=numTelefono, password=hashed_pw, verificato=False)
         else:
             raise ValueError("Ruolo non valido.")
 
@@ -118,9 +102,8 @@ class GestoreUtente:
         
         mail.send(msg)
     
-
     @staticmethod
-    def verficaEmail(token):
+    def verificaEmail(token):
         try:
             email = GestoreUtente.confermaTokenVerifica(token)
         except:
@@ -140,7 +123,6 @@ class GestoreUtente:
         utente.verificato = True
         db.session.commit()
         
-
     @staticmethod
     def generaTokenVerifica(email):
         #Inzializzo il serializer con la mia chiave privata
@@ -158,7 +140,6 @@ class GestoreUtente:
         except (SignatureExpired, BadTimeSignature):
             return None
         
-
     @staticmethod
     def modificaProfilo(utente, dati_form):        
         vecchia_password = dati_form.get('vecchia_password')
@@ -176,22 +157,43 @@ class GestoreUtente:
             # Genera il nuovo hash e aggiorna
             utente.password = generate_password_hash(nuova_password)
 
+
+        nome = dati_form.get('nome', utente.nome)
+        cognome = dati_form.get('cognome', utente.cognome)
+
+        if not nome or not cognome:
+            raise ValueError("Compila tutti i campi")
+
+
         # Aggiornamento dati anagrafici
-        utente.nome = dati_form.get('nome', utente.nome)
-        utente.cognome = dati_form.get('cognome', utente.cognome)
-        utente.numTelefono = dati_form.get('numTelefono', utente.numTelefono)
-        
+        utente.nome = nome
+        utente.cognome = cognome
+        utente.numTelefono =  dati_form.get('numTelefono', utente.numTelefono)
+
+
         # Se l'istanza dell'utente è uno Studente, aggiorna anche le informazioni accademiche
         if utente.ruolo == 'studente':
-            utente.universita = dati_form.get('universita', utente.universita)
-            utente.facolta = dati_form.get('facolta', utente.facolta)
-            utente.corso = dati_form.get('corso', utente.corso)
+            universita =dati_form.get('universita', utente.universita)
+            facolta = dati_form.get('facolta', utente.facolta)
+            corso = dati_form.get('corso', utente.corso)
+            if not universita or not facolta or not corso:
+                raise ValueError("Compila tutti i campi")
+
+
+            utente.universita = universita
+            utente.facolta = facolta
+            utente.corso = corso
+
+   
             
         db.session.commit()
 
     @staticmethod
     def eliminaProfilo(utente_id):
         utente = Utente.query.get(utente_id)
+        if not utente:
+            raise ValueError("Utente non trovato")
+        
         db.session.delete(utente)
         db.session.commit()
 
@@ -232,6 +234,8 @@ class GestoreUtente:
             #Aggiorno la password cifrata nel db solo dopo l'invio della email
             db.session.commit()
         except Exception:
+            #Se l'invio dell'eamil non va a buon fine riporto la sessione del database a quella iniziale e sollevo l'eccezione
+            db.session.rollback() 
             raise ValueError("Impossibile inviare l'email di recupero password.")
 
     @staticmethod
